+++ R/estimate_infections.R
@@ -295,6 +295,6 @@ estimate_infections <- function(data,
   )
 
   ## Join stan fit if required
-  class(ret) <- c("epinowfit", "estimate_infections", class(ret))
+  class(ret) <- c("estimate_infections", "epinowfit", class(ret))
   ret
 }